requirements.txt
yliveticker
protobuf==3.20.*
confluent-kafka
